rename : config, exception 폴더 common 하위 폴더로 이관
