+++ src/main/java/com/miracle/coordifit/common/config/HistoryConfig.java
@@ -1,4 +1,4 @@
-package com.miracle.coordifit.config;
+package com.miracle.coordifit.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
