+++ src/main/java/com/miracle/coordifit/user/service/UserService.java
@@ -11,8 +11,8 @@ import org.springframework.transaction.annotation.Transactional;
 import com.miracle.coordifit.auth.dto.AuthRequestDto;
 import com.miracle.coordifit.auth.dto.KakaoUserResponse;
 import com.miracle.coordifit.auth.service.IEmailService;
+import com.miracle.coordifit.common.exception.InactiveUserException;
 import com.miracle.coordifit.common.service.IFileService;
-import com.miracle.coordifit.exception.InactiveUserException;
 import com.miracle.coordifit.post.dto.PostDto;
 import com.miracle.coordifit.post.repository.PostRepository;
 import com.miracle.coordifit.user.dto.MyPageResponseDto;
